fix: sign-compare in the new pitch loop (-Werror caught it)
The pitch-honouring blit's int loop vars against the unsigned
FB_WIDTH/FB_HEIGHT broke the -Werror desktop build - the previous
commit was staged before its build check ran.  Cast the bounds.

+++ desktop/main.c
@@ -185,10 +185,10 @@ void plat_present(void)
         }
         /* row by row via the returned pitch - SDL does not guarantee
          * pitch == width * 4 */
-        for (y = 0; y < FB_HEIGHT; ++y) {
+        for (y = 0; y < (int)FB_HEIGHT; ++y) {
             uint32_t *row = (uint32_t *)(base + (size_t)y * (size_t)pitch);
             const uint8_t *src = fb + (size_t)y * FB_WIDTH;
-            for (x = 0; x < FB_WIDTH; ++x) {
+            for (x = 0; x < (int)FB_WIDTH; ++x) {
                 row[x] = src[x] ? on : 0xFF000000u;
             }
         }
